Adapted library build script to build fat binaries on OSX (ppc, i386, x86_64); updated example xcode project to include proper files and checked that it functions properly.

+++ scripts/build-libs.sh
@@ -1,11 +1,20 @@
 #!/bin/sh
 
+#NOTES
+# - to get fat binaries: http://www.mail-archive.com/libusbx-devel@lists.sourceforge.net/msg00442.html
+# - output logging (i.e., redirection) is not working properly for stderr and some commands (copying etc.)
+
 #initialize variables (the only things you should change are the ..._VERSION and ..._DL_URL variables)
 
 #NOTE: this patch removes a redefenition error in libusb-compat introduced by libusbx.
 #      according to this ticket (http://sourceforge.net/apps/trac/libusbx/ticket/31),
 #      this will be fixed in libusbx 1.0.13.
 COMPAT_X_PATCH=libusb-compat-0.1.4_libusbx.patch
+
+# *** Extra compiler flags
+ARCH_SPECS="-arch ppc -arch i386 -arch x86_64"
+EXTRA_CFLAGS=$ARCH_SPECS
+EXTRA_LDFLAGS=
 
 BASE_DIR=`pwd`/build-output
 BUILD_DIR=$BASE_DIR/build
@@ -40,10 +49,34 @@ function download_if_necessary() {
 
 
 
+### handle command-line arguments
+KEEP_INTERMEDIATES=no
+if [ $# -gt 0 ]; then
+	case $1 in
+	-h)
+		echo "This script is meant to compile the libraries needed by ofxGenericDmx."
+		echo "The following options can be used:"
+		echo "\t-h\tshow this help message"
+		echo "\t-k\tkeep all intermediate files"
+		exit 0
+		;;
+	-k)
+		KEEP_INTERMEDIATES=yes
+		;;
+	*)
+		echo "$0: Unknown arguments given, try running '$0 -h'"
+		exit 1
+		;;
+	esac
+fi
+
+
+
 ### prepare building
 mkdir -p $BASE_DIR
 cd $BASE_DIR
 mkdir -p $BUILD_DIR
+rm -f $LOG_FILE
 
 # detect OS
 OS_NAME=`uname -s`
@@ -66,10 +99,12 @@ tar xjf $USB_COMPAT_ARCH_NAME 2>&1 > /dev/null
 tar xzf $FTDI_ARCH_NAME 2>&1 > /dev/null
 echo >> $LOG_FILE
 
+cd $BASE_DIR
+
 ### build libusbx
 echo "*** Building library $USBX_BASE_NAME ***" 2>&1 | tee -a $LOG_FILE
 cd $USBX_BASE_NAME
-./configure --prefix=$BUILD_DIR --disable-shared 2>&1 >> $LOG_FILE
+CFLAGS="$CFLAGS $EXTRA_CFLAGS $ARCH_SPECS" LDFLAGS="$LDFLAGS $EXTRA_LDFLAGS" ./configure --prefix=$BUILD_DIR --disable-shared --disable-dependency-tracking 2>&1 >> $LOG_FILE
 make all install 2>&1 >> $LOG_FILE
 cd ..
 echo >> $LOG_FILE
@@ -78,7 +113,9 @@ echo >> $LOG_FILE
 echo "*** Building library $USB_COMPAT_BASE_NAME ***" 2>&1 | tee -a $LOG_FILE
 cd $USB_COMPAT_BASE_NAME
 patch -uNp1 < $BASE_DIR/../$COMPAT_X_PATCH 2>&1 >> $LOG_FILE
-LIBUSB_1_0_CFLAGS="-I$BUILD_DIR/include -I$BUILD_DIR/include/libusb-1.0" LIBUSB_1_0_LIBS=-L$BUILD_DIR/lib ./configure --prefix=$BUILD_DIR --disable-shared 2>&1 >> $LOG_FILE
+CFLAGS="$CFLAGS $EXTRA_CFLAGS" LDFLAGS="$LDFLAGS $EXTRA_LDFLAGS" \
+	LIBUSB_1_0_CFLAGS="-I$BUILD_DIR/include -I$BUILD_DIR/include/libusb-1.0" LIBUSB_1_0_LIBS=-L$BUILD_DIR/lib \
+	./configure --prefix=$BUILD_DIR --disable-shared --disable-dependency-tracking 2>&1 >> $LOG_FILE
 make all install 2>&1 >> $LOG_FILE
 cd ..
 echo >> $LOG_FILE
@@ -86,7 +123,8 @@ echo >> $LOG_FILE
 ### build libftdi
 echo "*** Building library $FTDI_BASE_NAME ***" 2>&1 | tee -a $LOG_FILE
 cd $FTDI_BASE_NAME
-CFLAGS=-I$BUILD_DIR/include ./configure --prefix=$BUILD_DIR --disable-shared --disable-libftdipp --without-examples --without-docs 2>&1 >> $LOG_FILE
+CFLAGS="$CFLAGS $EXTRA_CFLAGS -I$BUILD_DIR/include" LDFLAGS="$LDFLAGS $EXTRA_LDFLAGS" \
+	./configure --prefix=$BUILD_DIR --disable-shared --disable-dependency-tracking --disable-libftdipp --without-examples --without-docs 2>&1 >> $LOG_FILE
 make all install 2>&1 >> $LOG_FILE
 cd ..
 echo >> $LOG_FILE
@@ -110,10 +148,14 @@ echo >> $LOG_FILE
 
 ### clean up
 echo "*** Cleaning up ***" 2>&1 | tee -a $LOG_FILE
-rm -R $USBX_BASE_NAME
-rm -R $USB_COMPAT_BASE_NAME
-rm -R $FTDI_BASE_NAME
-rm -R $BUILD_DIR
+if [ "x$KEEP_INTERMEDIATES" = "xno" ]; then
+	rm -R $USBX_BASE_NAME
+	rm -R $USB_COMPAT_BASE_NAME
+	rm -R $FTDI_BASE_NAME
+	rm -R $BUILD_DIR
+else
+	echo "*** Leaving all intermediate files in place."
+fi
 echo >> $LOG_FILE
 
 ### inform user of further actions
